refactor(accounts): rename adapter.py to account.py
